+++ src/main/java/blackjack/domain/card/Denomination.java
@@ -1,0 +1,37 @@
+package blackjack.domain.card;
+
+public enum Denomination {
+    ACE("A", 11),
+    TWO("2", 2),
+    THREE("3", 3),
+    FOUR("4", 4),
+    FIVE("5", 5),
+    SIX("6", 6),
+    SEVEN("7", 7),
+    EIGHT("8", 8),
+    NINE("9", 9),
+    TEN("10", 10),
+    JACK("J", 10),
+    QUEEN("Q", 10),
+    KING("K", 10);
+
+    private final String name;
+    private final int score;
+
+    Denomination(final String name, final int score) {
+        this.score = score;
+        this.name = name;
+    }
+
+    public boolean isAce() {
+        return this == ACE;
+    }
+
+    public int getScore() {
+        return score;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
