+++ src/main/java/blackjack/domain/user/Dealer.java
@@ -1,11 +1,16 @@
-package blackjack.domain;
+package blackjack.domain.user;
 
 public class Dealer extends User {
     private static final String NAME = "딜러";
     private static final int DRAW_LIMIT = 16;
 
     @Override
+    public int getHitLimit() {
+        return 0;
+    }
+
+    @Override
     public String getName() {
-        return null;
+        return NAME;
     }
 }
